My changes

+++ demo/src/main/java/com/example/demo/Dao.java
@@ -4,8 +4,8 @@ import java.util.List;
 
 public interface Dao<E> {
 	void save(E e);
-	E load(long empid);
-	void delete(long empid);
+	E load(int empid);
+	void delete(int empid);
 	void update(E e);
 	List<E> loadAll();
